# openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus — Core Business Unified API (Nhom A6)
  version: 1.0.0
  contact:
    name: Nhom A6 Core Business
    email: support@campus.local
  description: |
    Hợp đồng API OpenAPI 3.1 hợp nhất cho nhóm A6 Core Business liên kết với các nhóm khác.
    Bao gồm các cặp đàm phán REST sync:
    - Pair 02: Core Business (A6) ↔ AI Vision (A4)
    - Pair 03: Core Business (A6) ↔ Access Gate (A3) [Core làm Consumer]
    - Pair 10: Access Gate (A3) ↔ Core Business (A6) [Core làm Provider]
servers:
  - url: http://localhost:4010
    description: Prism mock server
  - url: https://api.campus.local
    description: Class network placeholder
...
security:
  - bearerAuth: []
paths:
...
  /gates/{gateId}/status:
    get:
      operationId: getGateStatus
      summary: Lấy trạng thái hoạt động của cổng vật lý
      description: Truy xuất trạng thái vật lý thời gian thực hiện tại của một cổng kiểm soát (OPEN, CLOSED, LOCKED, FAULT).
      tags: [access-gate]
      parameters:
        - name: gateId
          in: path
          required: true
          description: Mã định danh duy nhất của cổng kiểm soát (định dạng GATE-XX).
          schema:
            type: string
            pattern: '^GATE-[0-9]{2}$'
      responses:
        '200':
          description: Lấy trạng thái cổng thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /access/check:
    post:
      operationId: checkAccessPolicy
      summary: Kiểm tra quyền ra vào realtime (Policy Engine)
      description: |
        Access Gate gọi trực tiếp tới Core Business khi có thẻ quẹt tại đầu đọc để thẩm định quyền ra/vào.
        API này yêu cầu tốc độ phản hồi cực kỳ nhanh (SLA < 50ms) để mở cổng ngay lập tức và tránh ùn tắc.
        Hỗ trợ chống trùng lặp yêu cầu thông qua requestId (idempotency).
      tags: [core-policy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessCheckRequest'
      responses:
        '200':
          description: Phản hồi quyết định thẩm định chính sách thành công (Cho phép hoặc Từ chối)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /policies/access/{policyId}:
    get:
      operationId: getAccessPolicyById
      summary: Lấy chi tiết chính sách ra vào
      description: Truy xuất thông tin cấu hình chi tiết của một chính sách phân quyền ra vào lớp học/tòa nhà.
      tags: [core-policy]
      parameters:
        - name: policyId
          in: path
          required: true
          description: Mã định danh chính sách (định dạng POL-XXX).
          schema:
            type: string
            pattern: '^POL-[0-9]{3}$'
      responses:
        '200':
          description: Lấy chi tiết chính sách thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessPolicy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /decisions/{decisionId}:
    get:
      operationId: getDecisionById
      summary: Tra cứu kết quả quyết định ra vào cụ thể
      description: Tra cứu dữ liệu chi tiết của một quyết định đã xử lý trước đó phục vụ hoạt động giám sát thời gian thực và kiểm toán bảo vệ.
      tags: [core-policy]
      parameters:
        - name: decisionId
          in: path
          required: true
          description: UUID định danh duy nhất của quyết định ra/vào.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lấy chi tiết quyết định thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessDecision'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    GateStatus:
      type: object
      additionalProperties: false
      required: [gateId, status, lastActivityAt, firmwareVersion]
      properties:
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Mã định danh của cổng vật lý.
          example: GATE-01
        status:
          type: string
          enum: [OPEN, CLOSED, LOCKED, FAULT]
          description: Trạng thái vật lý hiện tại của cổng.
          example: CLOSED
        lastActivityAt:
          type: string
          format: date-time
          description: Lần cuối cùng cổng ghi nhận hoạt động.
          example: '2026-06-01T10:05:00Z'
        firmwareVersion:
          type: string
          minLength: 1
          maxLength: 30
          description: Phiên bản firmware của bộ điều khiển cổng.
          example: gate-fw-v1.4.2
...
    AccessCheckRequest:
      type: object
      additionalProperties: false
      required: [requestId, cardId, gateId, direction, timestamp]
      properties:
        requestId:
          type: string
          format: uuid
          description: ID duy nhất của yêu cầu quẹt thẻ, dùng làm idempotency key.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2cafe
        cardId:
          type: string
          pattern: '^CARD-[0-9]{6}$'
          description: Mã thẻ được quẹt tại cổng.
          example: CARD-123456
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Mã cổng vật lý nơi quẹt thẻ.
          example: GATE-01
        direction:
          type: string
          enum: [IN, OUT]
          description: Hướng quẹt thẻ (Vào/Ra).
          example: IN
        timestamp:
          type: string
          format: date-time
          description: Thời gian quẹt thẻ vật lý ghi nhận tại đầu đọc.
          example: '2026-06-01T10:00:00Z'
    AccessCheckResponse:
      type: object
      additionalProperties: false
      required: [decisionId, allow, reasonCode, policyId, expiresAt]
      properties:
        decisionId:
          type: string
          format: uuid
          description: ID duy nhất của quyết định ra vào được sinh bởi Core.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2caff
        allow:
          type: boolean
          description: Quyết định cho phép mở cổng (true) hay từ chối (false).
          example: true
        reasonCode:
          type: string
          enum: [ALLOWED, INVALID_CARD, CARD_EXPIRED, GATE_LOCKED, TIME_RESTRICTION, DEPARTMENT_RESTRICTION, OUT_OF_SCHEDULE]
          description: Mã lý do giải thích cho quyết định allow/deny.
          example: ALLOWED
        policyId:
          type: [string, 'null']
          pattern: '^POL-[0-9]{3}$'
          description: Mã chính sách được áp dụng để đưa ra quyết định (nếu có, null nếu lỗi thẻ).
          example: POL-101
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: Thời điểm hết hạn của phiên cho phép quẹt thẻ đi qua (thường là 5 giây từ lúc allow để cổng tự đóng).
          example: '2026-06-01T10:00:05Z'
    AccessPolicy:
      type: object
      additionalProperties: false
      required: [policyId, name, description, department, status]
      properties:
        policyId:
          type: string
          pattern: '^POL-[0-9]{3}$'
          description: Mã định danh chính sách.
          example: POL-101
        name:
          type: string
          minLength: 3
          maxLength: 80
          description: Tên chính sách.
          example: Chính sách ra vào sinh viên chính quy CNTT
        description:
          type: string
          maxLength: 250
          description: Mô tả chi tiết chính sách.
          example: Cho phép sinh viên khoa CNTT ra vào cổng chính từ 07:00 đến 22:00 hàng ngày.
        department:
          type: string
          enum: [ALL, IT, BIZ, ENG, ARCH]
          description: Khoa/phòng ban áp dụng chính sách này.
          example: IT
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
          description: Trạng thái kích hoạt của chính sách.
          example: ACTIVE
    AccessDecision:
      type: object
      additionalProperties: false
      required: [decisionId, cardId, gateId, direction, timestamp, allow, reasonCode, evaluatedPolicyId]
      properties:
        decisionId:
          type: string
          format: uuid
          description: ID của quyết định ra vào.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2caff
        cardId:
          type: string
          pattern: '^CARD-[0-9]{6}$'
          description: Thẻ thực hiện quẹt.
          example: CARD-123456
        gateId:
          type: string
          pattern: '^GATE-[0-9]{2}$'
          description: Cổng quẹt thẻ.
          example: GATE-01
        direction:
          type: string
          enum: [IN, OUT]
          example: IN
        timestamp:
          type: string
          format: date-time
          example: '2026-06-01T10:00:00Z'
        allow:
          type: boolean
          example: true
        reasonCode:
          type: string
          enum: [ALLOWED, INVALID_CARD, CARD_EXPIRED, GATE_LOCKED, TIME_RESTRICTION, DEPARTMENT_RESTRICTION, OUT_OF_SCHEDULE]
          example: ALLOWED
        evaluatedPolicyId:
          type: [string, 'null']
          pattern: '^POL-[0-9]{3}$'
          example: POL-101
        operatorNote:
          type: [string, 'null']
          maxLength: 300
          description: Ghi chú thủ công từ admin/bảo vệ nếu cần can thiệp ghi đè.
          example: Khớp thẻ thành công.
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI định danh loại lỗi theo tiêu chuẩn RFC 7807.
          example: https://campus.local/errors/validation
        title:
          type: string
          minLength: 3
          maxLength: 120
          description: Tiêu đề lỗi ngắn gọn.
          example: Dữ liệu không hợp lệ
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code trả về.
          example: 400
        detail:
          type: [string, 'null']
          maxLength: 500
          description: Mô tả chi tiết nguyên nhân gây lỗi.
          example: Trường imageBase64 không đúng định dạng base64 hợp lệ
        instance:
          type: [string, 'null']
          format: uri
          description: URI của request cụ thể gây ra lỗi.
          example: https://campus.local/vision/detect
        errors:
          type: array
          minItems: 0
          maxItems: 50
          description: Danh sách lỗi chi tiết ở từng field cụ thể (nếu có).
          items:
            type: object
            additionalProperties: false
            required: [field, code, message]
            properties:
              field:
                type: string
                minLength: 1
                maxLength: 80
                example: imageBase64
              code:
                type: string
                minLength: 2
                maxLength: 80
                example: PATTERN_MISMATCH
              message:
                type: string
                minLength: 3
                maxLength: 200
                example: Dữ liệu ảnh base64 không đúng mẫu chuỗi base64
  responses:
    BadRequest:
      description: Dữ liệu gửi lên sai định dạng hoặc thiếu các trường bắt buộc
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/validation
            title: Dữ liệu không hợp lệ
            status: 400
            detail: Payload JSON gửi lên không đúng JSON Schema quy định
            instance: https://campus.local/vision/detect
            errors:
              - field: imageType
                code: ENUM_MISMATCH
                message: imageType bắt buộc phải là URL hoặc BASE64
    Unauthorized:
      description: Yêu cầu thiếu Bearer Token hoặc token không hợp lệ/hết hạn
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/unauthorized
            title: Chưa xác thực danh tính
            status: 401
            detail: Token JWT bị thiếu trong tiêu đề Authorization hoặc đã hết hạn sử dụng
            instance: https://campus.local/vision/detect
            errors: []
    Forbidden:
      description: Consumer không có quyền thực hiện thao tác phân tích này
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/forbidden
            title: Không được phép truy cập
            status: 403
            detail: Khóa API hoặc Token không chứa scope cần thiết (ai-vision:write)
            instance: https://campus.local/vision/detect
            errors: []
    NotFound:
      description: Không tìm thấy requestId hoặc model tương ứng trong hệ thống
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/not-found
            title: Không tìm thấy tài nguyên
            status: 404
            detail: Yêu cầu phân tích với requestId chỉ định không tồn tại hoặc đã bị xóa
            instance: https://campus.local/vision/detect/0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
            errors: []
...
    UnprocessableEntity:
      description: Dữ liệu đúng định dạng JSON Schema nhưng không xử lý được (ví dụ ảnh lỗi, hỏng, quá mờ)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/unprocessable-entity
            title: Lỗi xử lý nghiệp vụ hình ảnh
            status: 422
            detail: Không thể giải mã hoặc phân tích tệp ảnh gửi lên do tệp ảnh bị lỗi hoặc không hỗ trợ định dạng
            instance: https://campus.local/vision/detect
            errors: []
    InternalServerError:
      description: Lỗi hệ thống hoặc lỗi Runtime phía AI Vision Service không mong muốn
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://campus.local/errors/internal-server-error
            title: Lỗi hệ thống nội bộ
            status: 500
            detail: Gặp lỗi nghiêm trọng khi kết nối tới cụm xử lý Model AI (GPU cluster timeout)
            instance: https://campus.local/vision/detect
            errors: []